refactor: Moving plot_hdi_forest to mmm module, moving nrmse and nmae to metrics, removing MMMEvaluator and replacing with functions then updating the MLFlow module appropriately
